requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
cffi==1.16.0
characterai==1.0.1
charset-normalizer==3.3.2
curl_cffi==0.6.3
idna==3.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
typing_extensions==4.11.0
urllib3==2.2.1
vk-api==11.9.9
websockets==12.0
wheel==0.43.0
